requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.1
numpy==1.26.4
pandas==2.2.3
httpx==0.27.2
requests==2.32.3
openai==2.30.0
langchain==1.2.14
langgraph==1.1.4
langgraph-sdk==0.3.12
langgraph-prebuilt==1.0.8
langgraph-checkpoint==4.0.1
nebula3-python==3.8.3
PyMuPDF==1.21.0
pdfplumber==0.11.6
opencv-python==4.10.0.84
openpyxl==3.1.5
websockets==12.0
pymssql==2.3.2
anyio==4.12.1